+++ public-info/src/main/java/com/bank/publicinfo/dto/AuditDto.java
@@ -1,9 +1,8 @@
 package com.bank.publicinfo.dto;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
-public record AuditDto(BigInteger id,
+public record AuditDto(Long id,
                        String entityType,
                        String operationType,
                        String createdBy,
