+++ public-info/src/main/java/com/bank/publicinfo/repository/AuditRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface AuditRepository extends JpaRepository<Audit, BigInteger> {
+public interface AuditRepository extends JpaRepository<Audit, Long> {
 }
